FR-93 update components style
