--- src/styles/IdeasDonationProgressBar.css
@@ -1,49 +1,0 @@
-:root {
-    --primary-color: #1976d2;
-}
-
-.progress-bar-container {
-    max-width: 350px;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    border: 5px solid var(--primary-color);
-    border-radius: 25px;
-    padding: 10px;
-}
-
-.progress-bar-container .MuiLinearProgress-root {
-    width: 100%; /* Ensure the progress bar spans the entire container */
-    margin-bottom: 10px; /* Add spacing between the progress bar and other elements */
-    height: 20px;
-    border-radius: 25px;
-}
-
-.progress-bar-container .MuiTypography-root {
-    margin-bottom: 5px;
-    font-size: 20px;
-}
-
-.progress-bar-container .MuiTextField-root {
-    margin-bottom: 10px;
-    font-size: 20px;
-}
-
-.progress-bar-mark {
-    /* Add mark styles */
-}
-
-.percentage-text {
-    /* Add percentage text styles */
-}
-
-.small-card-container {
-    /* Add styles for small cards */
-}
-
-.large-card-container {
-    /* Add styles for large cards or pages */
-}
-
-@media (max-width: 576px) {
-}
