+++ src/styles/styles.css
@@ -1,5 +1,4 @@
 @import url('../styles/donation-form.css');
-@import url('../styles/IdeasDonationProgressBar.css');
 
 html {
     scroll-behavior: smooth;
